separate out pie chart tooltip
