+++ app/components/ChartPie.js
@@ -2,6 +2,8 @@ import { PieChart, Pie, Sector, Cell, Tooltip, Legend, ResponsiveContainer } fro
 var React = require('react');
 var PropTypes = React.PropTypes;
 
+import CustomToolTipPieChart from './CustomToolTipPieChart'
+
 import {
   BOX_BORDER,
   SPACING,
@@ -67,83 +69,6 @@ import {
   return color;
 }
 
-const tooltipstyle = {
-  width: '100%',
-  margin: 0,
-  lineHeight: 24,
-  border: '1px solid #f5f5f5',
-  backgroundColor: 'rgba(255, 255, 255, 0.95)',
-  padding: 10,
-};
-
-const toolTipLabel = {
-  margin: '0',
-  color: '#666',
-  fontWeight: '700',
-};
-
-const CustomTooltip  = React.createClass({
-  propTypes: {
-    type: PropTypes.string,
-    payload: PropTypes.array,
-    label: PropTypes.string,
-  },
-
-  render() {
-    const { active } = this.props;
-    let html_hov = '';
-    if (active) {
-      const { payload, label } = this.props;
-
-
-      const values = this.props.data.map( val => {
-        return val.value
-      })
-      //get the total of the values
-      var total = values.reduce(function(a, b) {
-        return a + b;
-      }, 0);
-
-      const thedata = payload.map( bar_segment => {
-
-       const colors = get_keyColors(bar_segment.name)
-        const toolTipName = {
-          margin: '0',
-          color: '#000',
-        }
-
-        const toolTipValue = {
-          fontWeight: '800',
-        }
-
-
-        const value = bar_segment.value ? ((bar_segment.value/total) * 100).toFixed(1).toString() + '%' : '';
-        const name = bar_segment.name + ": "
-
-        return ( <p key={bar_segment.name} style={toolTipName}>{name}<span style={toolTipValue}>{value}</span></p>)
-      })
-
-      let tooltip = (<div />)
-      if (label === '1' || label === '2' ){
-        return (<div key={label+'blanktip'}/>)
-      }
-
-      const labelstr = label.toString();
-
-      return (
-
-        <div key={labelstr+'tooltip'} style={tooltipstyle}>
-          {thedata}
-      </div>
-
-      );
-    }
-
-    return null;
-  }
-});
-
-
 //custom legend. to display landuse landcover category with correct fill
 const renderLegendPercent = (props) => {
   const { payload } = props;
@@ -165,14 +90,12 @@ const renderLegendPercent = (props) => {
     <div className="ui list">
       {
         payload.map((entry, index) => (
-
           <div className="item"  key={`item-${index}`}>
-
-          <svg  width="14" height="14" >
-            <path stroke-strokeWidth="4" fill={entry.fill} stroke={entry.stroke} d="M0,0h32v32h-32z" >
-            </path>
-          </svg>
-          {`  ${entry.name}  (${((entry.value/total) * 100).toFixed(1)}%)` }
+            <svg  width="14" height="14" >
+              <path stroke-strokeWidth="4" fill={entry.fill} stroke={entry.stroke} d="M0,0h32v32h-32z" >
+              </path>
+            </svg>
+            {`  ${entry.name}  (${((entry.value/total) * 100).toFixed(1)}%)` }
           </div>
         ))
       }
@@ -277,17 +200,13 @@ const ChartPie = React.createClass({
                     data.map((entry, index) => <Cell key={index} fill={get_keyColors(entry.name)}/>)
                   }
                 </Pie>
-                <Tooltip content={<CustomTooltip data={data}/>}/>
+                <Tooltip content={<CustomToolTipPieChart get_keyColors={get_keyColors} data={data}/>}/>
               </PieChart>
 
             }
           </div>
         </div>
       </div>
-
-
-
-
 
     );
   }
