+++ app/components/CustomToolTipSimpleBarCharts.js
@@ -24,20 +24,21 @@ const CustomToolTipSimpleBarCharts  = React.createClass({
     payload: PropTypes.array,
     label: PropTypes.string,
   },
-
   render() {
+
     const { active } = this.props;
     let html_hov = '';
     if (active) {
       const { payload, label } = this.props;
 
-      const thedata = payload.map( bar_segment => {
+      const reversed_payload = [ ...payload ].reverse()
 
-        const colors = this.props.get_keyColors(bar_segment.name)
+      const thedata = reversed_payload.map( bar_segment => {
+       const colors = this.props.get_keyColors(bar_segment.name)
 
         const toolTipName = {
           margin: '0',
-          color: colors,
+          color: colors[1],
         }
 
         const toolTipValue = {
@@ -47,22 +48,35 @@ const CustomToolTipSimpleBarCharts  = React.createClass({
         const value = bar_segment.value ?  bar_segment.value.toString().substring(0,5) : 'N/A';
         const name = bar_segment.name + ": "
 
-        return ( <p key={bar_segment.name} style={toolTipName}>{name}<span style={toolTipValue}>{value}</span></p>)
+        //when tra's have a value of 0 do not display the tool tip...
+        if((bar_segment.value === 0 || !bar_segment.value ) && this.props.chart_type.toUpperCase() === 'TRA'){
+          return null
+        } else {
+          return ( <p key={bar_segment.name} style={toolTipName}>{name}<span style={toolTipValue}>{value}</span></p>)
+        }
       })
 
-      if (label === '1' || label === '2' ){
-        return (<div key={label+'blanktip'}/>)
+      //check to see if there is data in the case of tra's there could be no data and
+      //  we want to tell users that there is no data.
+      let hasdata = false;
+      thedata.map( d  => {
+        if(d){hasdata = true}
+      })
+
+      const labelstr = label.toString().trim();
+
+      //null tip when there is no id
+      if (!labelstr || !hasdata){
+        return (<div key={labelstr+'blanktip'} />)
       }
 
-      const labelstr = label.toString();
 
+      //return tooltip
       return (
-
         <div key={labelstr+'tooltip'} style={tooltipstyle}>
-          <p style={toolTipLabel}>Cathcment: {`${labelstr}`}</p>
+          <p style={toolTipLabel}>{this.props.level_label}: {`${labelstr}`}</p>
           {thedata}
-      </div>
-
+        </div>
       );
     }
